modify some view

+++ controllers/api/SamplesController.php
@@ -16,7 +16,9 @@ class SamplesController extends DefaultController {
 	}
 	
 	public function actionView($id){
-
+	
+		if (!yii::$app->user->can('sample-samples-view'))
+			return ['status' => 403 , 'message' => 'Unauthorize'];
 		
 		$model = Samples::findOne($id);
 		
@@ -27,10 +29,17 @@ class SamplesController extends DefaultController {
 	}
 	
 	public function actionIndex(){
+		
+		if (!Yii::$app->user->can('sample-samples-manage'))
+			return ['status' => 403 , 'message' => 'Unauthorize'];
+
 		return Samples::find()->all();
 	}
 	
 	public function actionCreate(){
+		
+		if (!Yii::$app->user->can('sample-samples-create'))
+			return ['status' => 403 , 'message' => 'Unauthorize'];
 		
 		$postParams = yii::$app->request->post();
 		
@@ -50,6 +59,9 @@ class SamplesController extends DefaultController {
 	
 	public function actionUpdate($id){
 		
+		if (!Yii::$app->user->can('sample-samples-update'))
+			return ['status' => 403 , 'message' => 'Unauthorize'];
+		
 		$postParams = yii::$app->request->post();
 		
 		$model = Samples::findOne($id);
@@ -64,6 +76,9 @@ class SamplesController extends DefaultController {
 	}
 	
 	public function actionDelete($id){
+		
+		if (!Yii::$app->user->can('sample-samples-delete'))
+			return ['status' => 403 , 'message' => 'Unauthorize'];
 		
 		if(Samples::findOne($id)->delete())
 			return ['status' => 1];
